third_party/blink/web_tests/fast/forms/form/move-option-between-documents-expected.txt

PASS select.options.namedItem('option') is null
PASS successfullyParsed is true

TEST COMPLETE
